Move postgres singleton code to class
Classes are easier to test and nicer to organize than singletons. The
calling interface and tests remain the same for now, but those will
both change in future commits. The purpose of this commit is to
refactor the logic of the withDb function to a class method.
